Do an inbound ink eye animation.

+++ led_display.h
@@ -56,7 +56,7 @@ void do_brightness_correction();
 
 extern uint16_t face_banks[4];
 extern rgbcolor_t leg_colors_curr[8];
-extern uint16_t face_ambient_brightness;
+extern uint32_t face_ambient_brightness;
 extern uint8_t face_state;
 extern uint64_t face_ambient;
 extern uint8_t wiggle_mask;
